websocket/requirements.txt
channels
channels-redis
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers